Restructure graphics library headers.
