+++ src/Graphics/include/litefx/gfx/blit.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <litefx/graphics_api.hpp>
+#include <litefx/math.hpp>
+
+namespace LiteFX::Graphics {
+    using namespace LiteFX;
+
+}
