--- src/Graphics/include/litefx/graphics_formatters.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <format>
-#include "graphics_api.hpp"
-
-using namespace LiteFX::Graphics;
